Moving NewRequest into events package
